requirements.txt
azure-cognitiveservices-speech
python-dotenv
flask
flask-socketio